backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.1
